requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
aiosqlite==0.19.0
